refactor: move voices.py to cli module
